fix(memory-core): keep legacy config migration under doctor
